// ts-node/model/eventEventMessageMaintenance.ts
/**
 * Moira Alert
 * This is an API description for Moira Alert project. Please check https://github.com/moira-alert
 *
 * The version of the OpenAPI document: 2.5.1.47
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';

export class EventEventMessageMaintenance {
    'setupUser'?: string;
    'setupTime'?: number;
    'removeUser'?: string;
    'removeTime'?: number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "setupUser",
            "baseName": "setup_user",
            "type": "string"
        },
        {
            "name": "setupTime",
            "baseName": "setup_time",
            "type": "number"
        },
        {
            "name": "removeUser",
            "baseName": "remove_user",
            "type": "string"
        },
        {
            "name": "removeTime",
            "baseName": "remove_time",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return EventEventMessageMaintenance.attributeTypeMap;
    }
}