// ts-node/model/inlineResponse200.ts
/**
 * Moira Alert
 * This is an API description for Moira Alert project. Please check https://github.com/moira-alert
 *
 * The version of the OpenAPI document: 2.5.1.47
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { InlineResponse200Contacts } from './inlineResponse200Contacts';

export class InlineResponse200 {
    /**
    * Flag for determining if Moira is accessible remotely.
    */
    'remoteAllowed'?: boolean;
    /**
    * List of enabled contact types
    */
    'contacts'?: Array<InlineResponse200Contacts>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "remoteAllowed",
            "baseName": "remoteAllowed",
            "type": "boolean"
        },
        {
            "name": "contacts",
            "baseName": "contacts",
            "type": "Array<InlineResponse200Contacts>"
        }    ];

    static getAttributeTypeMap() {
        return InlineResponse200.attributeTypeMap;
    }
}