// ts-node/model/inlineResponse2003List.ts
/**
 * Moira Alert
 * This is an API description for Moira Alert project. Please check https://github.com/moira-alert
 *
 * The version of the OpenAPI document: 2.5.1.47
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from '../api';
import { Trigger } from './trigger';

export class InlineResponse2003List {
    'metrics'?: Array<string>;
    'pattern'?: string;
    'triggers'?: Array<Trigger>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "metrics",
            "baseName": "metrics",
            "type": "Array<string>"
        },
        {
            "name": "pattern",
            "baseName": "pattern",
            "type": "string"
        },
        {
            "name": "triggers",
            "baseName": "triggers",
            "type": "Array<Trigger>"
        }    ];

    static getAttributeTypeMap() {
        return InlineResponse2003List.attributeTypeMap;
    }
}